Documentation

BiologicalOscillations.gene_regulatory_network — Method

gene_regulatory_network(connectivity::AbstractMatrix)

Creates a ReactionSystem of interacting mRNA and proteins based on the provided connectivity.

Arguments

connectivity::AbstractMatrix: A 2 dimensional matrix filled with -1, 0, and 1 values indicating the edges of the network.

BiologicalOscillations.grn_parameters — Method

grn_parameters(model:ReactionSystem, α::AbstractVector, β::AbstractVector, δ::AbstractVector, γ::AbstractVector, κ::AbstractVector, η::AbstractVector)

Creates an ordered parameter vector for a model created with the gene_regulatory_network function to use in ODEProblem.

Arguments

model::ReactionSystem: Model generated with protein_interaction_network

α:AbstractVector: Vector of intrinsic mRNA degradation rates

β:AbstractVector: Vector of intrinsic protein synthesis rates

δ:AbstractVector: Vector of intrinsic protein degradation rates

γ:AbstractVector: Vector of strengths for each network interaction

κ:AbstractVector: Vector of midpoints for each network interaction

η:AbstractVector: Vector of sensitivity values for each network interaction

Note

It is assumed that parameters follow the same order as the connectivity matrix. Namely, the first node is encoded on the first row of the connectivty matrix and the first edge comes from the first nonzero element of the connectivity.

BiologicalOscillations.grn_timescale — Method

grn_timescale(α::AbstractVector, δ::AbstractVector)

Calculate the slowest timescale of the gene regulatory network

Arguments

α:AbstractVector: Vector of intrinsic mRNA degradation rates

δ:AbstractVector: Vector of intrinsic protein degradation rates

Returns

timescale::Real