Elowitz & Leibler 2000 - Repressilator

BiologicalOscillations.elowitz_2000 — Constant

elowitz_2000

Generalized Repressilator model based on Elowitz and Leibler 2000 Nature article.

In contrast to the original model, this implementation has independent parameters for each mRNA and protein pair:

odesys = convert(ODESystem, elowitz_2000) # hide
eq = latexify(odesys) # hide
LaTeXEquation(eq) # hide

where

\[\text{hillr}(P(t), \alpha, K, n) = \alpha \frac{K^n}{P(t)^n + K^n}\]

Using parameters

\[\begin{aligned} \alpha_1 = \alpha_2 = \alpha_3 &= 0.5 \\ \beta_1 = \beta_2 = \beta_3 &= 0.05 \\ \gamma_1 = \gamma_2 = \gamma_3 &= 0.005 \\ \delta_1 = \delta_2 = \delta_3 &= 0.0025 \\ K_1 = K_2 = K_3 &= 40 \\ n_1 = n_2 = n_3 &= 2 \\ \mu_1 = \mu_2 = \mu_3 &= 0.005 \end{aligned}\]

Under the initial condition

\[\begin{aligned} m_1(0) = m_2(0) = m_3(0) &= 0.0 \\ P_1(0) &= 20.0 \\ P_2(0) = P_3(0) &= 0.0 \end{aligned}\]

we obtain:

plot(sol) # hide