Guan et al. 2008 - Embryonic cell cycles

BiologicalOscillations.guan_2008 — Constant

guan_2008

Embryonic cell cycle model based on Guan et al. 2008 eLife article.

The article A robust and tunable mitotic oscillator in artificial cells describes the development of an artificial cell that sustains cell cycle signaling oscillations. In their work, the authors develop a mathematical model to describe the activity of two key proteins driving these oscillations: Cyclin B1 and Cdk1. Using a two-ODE model, the authors can capture the essential behavior of their artificial cell setup. Here, we base model used in their study under through the ReactionNetwork called guan_2008. The equations governing the evolution of the system are:

\[ \begin{align*} \frac{\mathrm{d} B\left( t \right)}{\mathrm{d}t} =& k_{S} + \left( - a_{D} - \mathrm{hill}\left( \left|C\left( t \right)\right|, b_{D}, K_{D}, n_{D} \right) \right) B\left( t \right) \\ \frac{\mathrm{d} C\left( t \right)}{\mathrm{d}t} =& k_{S} + \frac{\left( - a_{T} - \mathrm{hill}\left( \left|C\left( t \right)\right|, b_{T}, K_{T}, n_{T} \right) \right) C\left( t \right)}{\sqrt{r}} + \frac{\left( a_{T} + \mathrm{hill}\left( \left|C\left( t \right)\right|, b_{T}, K_{T}, n_{T} \right) \right) B\left( t \right)}{\sqrt{r}} + \left( - a_{D} - \mathrm{hill}\left( \left|C\left( t \right)\right|, b_{D}, K_{D}, n_{D} \right) \right) C\left( t \right) + \left( - a_{W} - \mathrm{hillr}\left( \left|C\left( t \right)\right|, b_{W}, K_{W}, n_{W} \right) \right) \sqrt{r} C\left( t \right) \end{align*} \]

Using the parameters specified in the manuscript (with r=1.5) we obtain the following solution: