Documentation

BiologicalOscillations.pin_parameters — Method

pin_parameters(model:ReactionSystem, α::AbstractVector, β::AbstractVector, γ::AbstractVector, κ::AbstractVector, η::AbstractVector)

Creates a parameter vector to use in ODEProblem in the appropriate order. Values provided in α, β, γ, κ, and η are mapped to their corresponding counterparts in the provided model.

Arguments

model::ReactionSystem: Model generated with protein_interaction_network

α:AbstractVector: Vector of intrinsic protein activation rates for each node

β:AbstractVector: Vector of intrinsic protein deactivation rates for each node

γ:AbstractVector: Vector of strengths for each network interaction

κ:AbstractVector: Vector of midpoints for each network interaction

η:AbstractVector: Vector of sensitivity values for each network interaction

Note

It is assumed that α and β's order follows the same order as the connectivity matrix that defined the model. Similarly, the order in γ, κ, and η follows the order in which edges appear in the connectivity matrix.

BiologicalOscillations.protein_interaction_network — Method

protein_interaction_network(connectivity::AbstractMatrix)

Creates a ReactionSystem of interacting proteins based on the provided connectivity.

Arguments

connectivity::AbstractMatrix: A 2 dimensional matrix filled with -1, 0, and 1 values indicating the edges of the network.