Documentation

BiologicalOscillations.find_grn_oscillations — Method

find_grn_oscillations(connectivity::AbstractMatrix, samples::Int; initial_conditions::AbstractVector=NaN)

Find oscillatory parameter sets in a gene regulatory network.

Arguments (Required)

connectivity::AbstractMatrix: Connectivity matrix for the gene regulatory network

samples::Int: Number of parameter sets to test

Arguments (Optional)

hyperparameters::Dict: Dictionary of hyperparameters for the algorithm. Default values are defined in DEFAULT_GRN_HYPERPARAMETERS.

Returns

pin_result::Dict: A ditionary containing the results of the oscillatory parameter set search. Output is encoded as:

pin_result = Dict("model" => "ReactionSystem of the gene regulatory network",
                  "parameter_sets" => "Dataframe of parameter sets",
                  "equilibration_result" => "Dataframe containing result metrics for the equilibration of parameter sets",
                  "simulation_result" => "Dataframe containing result metrics for the simulated parameter sets",)

BiologicalOscillations.gene_regulatory_network — Method

gene_regulatory_network(connectivity::AbstractMatrix)

Creates a ReactionSystem of interacting mRNA and proteins based on the provided connectivity.

Arguments

connectivity::AbstractMatrix: A 2 dimensional matrix filled with -1, 0, and 1 values indicating the edges of the network.

BiologicalOscillations.grn_equilibration_times — Method

grn_equilibration_times(model::ReactionSystem, parameter_sets::AbstractArray; equilibration_time_multiplier=10)

Calculate the equilibration times for a group of parameter sets.

Arguments (Required)

model::ReactionSystem: Model generated with gene_regulatory_network

parameter_sets::AbstractArray: Parameter sets generated with grn_parameter_sets

Arguments (Optional)

equilibration_time_multiplier::Real=10: Multiplier for the equilibration time. The equilibration time is calculated as the timescale of the system multiplied by this value.

Returns

equilibration_times::AbstractVector: Equilibration times for each parameter set.

BiologicalOscillations.grn_hit_rate — Method

grn_hit_rate(connectivity::AbstractMatrix, initial_samples::Int; target_oscillations::Int=100, max_samples::Int=1000000, max_trials::Int=5, hyperparameters=DEFAULT_GRN_HYPERPARAMETERS)

Estimates the hit rate of the oscillatory parameter set search in a gene regulatory network.

Arguments (Required)

connectivity::AbstractMatrix: Connectivity matrix of the gene regulatory network

initial_samples::Int: Number of parameter sets to sample in the first trial

Arguments (Optional)

target_oscillations::Int: Target number of oscillatory parameter sets

max_samples::Int: Maximum number of samples allowed

max_trials::Int: Maximum number of trials allowed

hyperparameters::Dict: Dictionary of hyperparameters for the algorithm. Default values are defined in DEFAULT_GRN_HYPERPARAMETERS.

Returns

hit_rate::Real: Estimated hit rate

BiologicalOscillations.grn_nodes_edges — Method

grn_nodes_edges(model::ReactionSystem)

Returns the number of nodes and edges of a model created with gene_regulatory_network.

Arguments

model::ReactionSystem: Model generated with gene_regulatory_network

Returns

AbstractVector: Array with 2 elements representing the number of nodes and edges

BiologicalOscillations.grn_parameter_sets — Method

grn_parameter_sets(model::ReactionSystem, samples::Int; dimensionless_time=true, parameter_limits=DEFAULT_GRN_PARAMETER_LIMITS, sampling_scales=DEFAULT_GRN_SAMPLING_SCALES, sampling_style="lhc")

Generate a set of parameter values for a gene regulatory network model.

Arguments (Required)

model::ReactionSystem: Model generated with gene_regulatory_network

samples::Int: Number of parameter sets to generate

random_seed::Int: Random seed for the sampling algorithm

Arguments (Optional)

dimensionless_time::Bool=true: If true, α₁ is set to 1.0 as well as the first N κ's with N=number of nodes. This is done to make the timescale of the system dimensionless.

parameter_limits::Dict: Dict with keys "α", "β", "δ", "γ", "κ", "η" and values of tuples with lower and upper limits for each parameter. For default value see DEFAULT_GRN_PARAMETER_LIMITS.

sampling_scales::Dict: Dict with keys "α", "β", "δ", "γ", "κ", "η" and values of strings with the sampling scale for each parameter. For default value see DEFAULT_GRN_SAMPLING_SCALES.

sampling_style::String="lhc": Sampling style. Options are "lhc" for latin hypercube sampling and "random" for random sampling.

BiologicalOscillations.grn_parameters — Method

grn_parameters(model:ReactionSystem, α::AbstractVector, β::AbstractVector, δ::AbstractVector, γ::AbstractVector, κ::AbstractVector, η::AbstractVector)

Creates an ordered parameter vector for a model created with the gene_regulatory_network function to use in ODEProblem.

Arguments

model::ReactionSystem: Model generated with gene_regulatory_network

α:AbstractVector: Vector of intrinsic mRNA degradation rates

β:AbstractVector: Vector of intrinsic protein synthesis rates

δ:AbstractVector: Vector of intrinsic protein degradation rates

γ:AbstractVector: Vector of strengths for each network interaction

κ:AbstractVector: Vector of midpoints for each network interaction

η:AbstractVector: Vector of sensitivity values for each network interaction

Note

It is assumed that parameters follow the same order as the connectivity matrix. Namely, the first node is encoded on the first row of the connectivty matrix and the first edge comes from the first nonzero element of the connectivity.

BiologicalOscillations.grn_timescale — Method

grn_timescale(α::AbstractVector, δ::AbstractVector)

Calculate the slowest timescale of the gene regulatory network

Arguments

α:AbstractVector: Vector of intrinsic mRNA degradation rates

δ:AbstractVector: Vector of intrinsic protein degradation rates

Returns

timescale::Real