Elowitz & Leibler 2000 - Repressilator

BiologicalOscillations.elowitz_2000 — Constant

elowitz_2000

Generalized Repressilator model based on Elowitz and Leibler 2000 Nature article.

In contrast to the original model, this implementation has independent parameters for each mRNA and protein pair:

\[ \begin{align*} \frac{\mathrm{d} m_1\left( t \right)}{\mathrm{d}t} =& \mathrm{hillr}\left( P_3\left( t \right), \alpha_3, K_3, n_3 \right) + \gamma_1 - m_1\left( t \right) \delta_1 \\ \frac{\mathrm{d} m_2\left( t \right)}{\mathrm{d}t} =& \mathrm{hillr}\left( P_1\left( t \right), \alpha_1, K_1, n_1 \right) + \gamma_2 - m_2\left( t \right) \delta_2 \\ \frac{\mathrm{d} m_3\left( t \right)}{\mathrm{d}t} =& \mathrm{hillr}\left( P_2\left( t \right), \alpha_2, K_2, n_2 \right) + \gamma_3 - m_3\left( t \right) \delta_3 \\ \frac{\mathrm{d} P_1\left( t \right)}{\mathrm{d}t} =& - P_1\left( t \right) \mu_1 + m_1\left( t \right) \beta_1 \\ \frac{\mathrm{d} P_2\left( t \right)}{\mathrm{d}t} =& - P_2\left( t \right) \mu_2 + m_2\left( t \right) \beta_2 \\ \frac{\mathrm{d} P_3\left( t \right)}{\mathrm{d}t} =& m_3\left( t \right) \beta_3 - P_3\left( t \right) \mu_3 \end{align*} \]

where

\[\text{hillr}(P(t), \alpha, K, n) = \alpha \frac{K^n}{P(t)^n + K^n}\]

We can obtain an oscillatory solution by using the following parameters

$\alpha_1, \alpha_2, \alpha_3$	0.5	$K_1, K_2, K_3$	40	$\beta_1, \beta_2, \beta_3$	0.05	$n_1, n_2, n_3$	2

$\gamma_1, \gamma_2, \gamma_3$	0.005	$\mu_1, \mu_2, \mu_3$	0.005	$\delta_1, \delta_2, \delta_3$	0.0025

and the initial condition

$m_1(0), m_2(0), m_3(0)$	0.0	$P_1(0)$	20.0	$P_2(0), P_3(0)$	0.0