Documentation

BiologicalOscillations.pin_parameters — Method

pin_parameters(model:ReactionSystem, α::AbstractVector, β::AbstractVector, γ::AbstractVector, κ::AbstractVector, η::AbstractVector)

Creates an ordered parameter vector for a model created with the protein_interaction_network function to use in ODEProblem.

Arguments

model::ReactionSystem: Model generated with protein_interaction_network

α:AbstractVector: Vector of intrinsic protein activation rates for each node

β:AbstractVector: Vector of intrinsic protein deactivation rates for each node

γ:AbstractVector: Vector of strengths for each network interaction

κ:AbstractVector: Vector of midpoints for each network interaction

η:AbstractVector: Vector of sensitivity values for each network interaction

Note

It is assumed that parameters follow the same order as the connectivity matrix. Namely, the first node is encoded on the first row of the connectivty matrix and the first edge comes from the first nonzero element of the connectivity.

BiologicalOscillations.pin_timescale — Method

pin_timescale(α::AbstractVector, β::AbstractVector, γ::AbstractVector)

Calculate the slowest timescale of the protein interaction network

Arguments

α:AbstractVector: Vector of intrinsic protein activation rates for each node

β:AbstractVector: Vector of intrinsic protein deactivation rates for each node

γ:AbstractVector: Vector of strengths for each network interaction

Returns

timescale::Real

BiologicalOscillations.protein_interaction_network — Method

protein_interaction_network(connectivity::AbstractMatrix)

Creates a ReactionSystem of interacting proteins based on the provided connectivity.

Arguments

connectivity::AbstractMatrix: A 2 dimensional matrix filled with -1, 0, and 1 values indicating the edges of the network.